Py3k support.
The changes required for python 3 support were actally relatively
minor. First, the exception syntax (`except Exception, e` vs
`except Exception as e`) and a couple of spots where `bytes` were
used when `str` was expected. Namely, in `MantaClient._request()`
the authorization header included a signature that was `bytes` but
needed to be cast as `str` and in the cases where a JSON response
needs to be parsed the `bytes` also need to be cast as `str`.

Changes tested in Python 2.7 and 3.4.